base_flask/requirements.txt
Flask
Flask-RESTx
flask_jwt_extended
Flask-SQLAlchemy
psycopg2-binary==2.9.7 
python-dotenv 
requests
flask-migrate
